Search patient page and most of search patient controller
The search page takes a patient_number (not id), if the patient is
found, it redirects to the view page and displays all of the patient
information in a string, if the patient is not found, it displays a
message saying it coudln't find it.
I still need to fix this 'patient not found' message so it gives the
user an option to Add a patient.
I did most of the controller, but it may not be done correctly so this
can be changed or tweaked as necessary.